requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
docstring_parser==0.16
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
grpcio==1.62.2
idna==3.7
imageio==2.34.0
importlib_resources==6.4.0
Jinja2==3.1.3
jsonargparse==4.28.0
kornia==0.7.2
kornia_rs==0.1.3
lightning==2.2.1
lightning-utilities==0.11.2
Markdown==3.6
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
plyfile==1.0.3
protobuf==5.26.1
pytorch-lightning==2.2.3
PyYAML==6.0.1
scipy==1.12.0
six==1.16.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
torch==2.2.1
torchaudio==2.2.1
torchmetrics==1.3.2
torchvision==0.17.1
tqdm==4.66.2
triton==2.2.0
typeshed_client==2.5.1
typing_extensions==4.11.0
Werkzeug==3.0.2
yarl==1.9.4
